% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iota_version_2.R
\name{check_conformity}
\alias{check_conformity}
\title{Check assumptions of weak superiority}
\usage{
check_conformity(aem)
}
\arguments{
\item{aem}{matrix of probabilities}
}
\value{
Returns the number of violations of the assumption of weak superiority.
0 if the assumptions are fulfilled.
}
\description{
This function tests if the probabilities within the AEM-matrix are in line with
the assumption of weak superiority
}
