% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_conformity_c}
\alias{check_conformity_c}
\title{Check assumptions of weak superiority}
\usage{
check_conformity_c(aem)
}
\arguments{
\item{aem}{matrix of probabilities}
}
\value{
Returns the number of violations of the assumption of weak superiority.
0 if the assumptions are fulfilled.
}
\description{
This function tests if the probabilities within the Assignment Error Matrix
are in line with the assumption of weak superiority.
}
\references{
Berding, Florian, and Pargmann, Julia (2022).Iota Reliability Concept
of the Second Generation.Measures for Content Analysis Done by
Humans or Artificial Intelligences. Berlin: Logos.
https://doi.org/10.30819/5581
}
