% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{est_expected_categories}
\alias{est_expected_categories}
\title{Estimate Expected Categories}
\usage{
est_expected_categories(data, aem)
}
\arguments{
\item{data}{\code{Matrix} which contains the codings for every coding unit. The
coding units must be in the rows and the raters must be in the columns. At
least two raters are necessary.}

\item{aem}{Assignment Error Matrix based on Iota Concept generation two (Iota2).}
}
\value{
Returns a \code{matrix} with the original data, the conditioned
probability each true category, and the expected category for every coding unit.
}
\description{
Function for estimating the expected category of coding units.
}
