% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iota_version_2.R
\name{get_patterns_unordered}
\alias{get_patterns_unordered}
\title{Get patterns unordered}
\usage{
get_patterns_unordered(obs_pattern_shape, categorical_levels)
}
\arguments{
\item{obs_pattern_shape}{\code{Matrix} generated by the function
\code{\link{get_patterns}} containing the observed patterns.}

\item{categorical_levels}{\code{Vector} containing all possible categories of
the content analysis.}
}
\value{
Function returns a \code{List} with the following components:
\item{n_unordered}{Number of unordered patterns.}
\item{shape_unordered}{\code{Matrix} which contains the unique unordered
patterns.}
}
\description{
Auxiliary function written in \code{R} for providing the necessary information
about the patterns generated by raters.This function does not produce the
input for the EM-algorithm. It is used to check for the identification
of the estimated model
}
