% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_random_start_values_class_sizes}
\alias{get_random_start_values_class_sizes}
\title{Generating randomly chosen probabilities for categorical sizes}
\usage{
get_random_start_values_class_sizes(n_categories)
}
\arguments{
\item{n_categories}{Integer for the number of categories in the data. Must be
at least 2.}
}
\value{
Returns a vector of randomly chosen categorical sizes.
}
\description{
Function written in \code{C++} for generating a set of randomly chosen
probabilities describing the size of the different classes. The
probabilities describe the relative frequencies of the categories in the data.
}
