% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_random_start_values_p}
\alias{get_random_start_values_p}
\title{Generating randomly chosen probabilities for Assignment Error Matrix}
\usage{
get_random_start_values_p(n_categories)
}
\arguments{
\item{n_categories}{Integer for the number of categories in the data. Must be
at least 2.}
}
\value{
Returns a matrix for Assignment Error Matrix (AEM) with randomly
generated probabilities. The generated probabilities are in line with
the assumption of weak superiority.
}
\description{
Function written in \code{C++} for generating a set of randomly chosen
probabilities for the Assignment Error Matrix.
}
