% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{get_summary}
\alias{get_summary}
\title{Get Summary}
\usage{
get_summary(object)
}
\arguments{
\item{object}{An object of class \code{iotarelr_iota2} created by
\code{\link{compute_iota2}}, \code{\link{check_new_rater}}, or
\code{\link{check_dgf}}.}
}
\value{
Prints central statistics of the estimated model.
}
\description{
Function for creating a short summary of the estimated Iota components.
}
