% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_ll}
\alias{grad_ll}
\title{Gradient for Log Likelihood in Condition Stage}
\usage{
grad_ll(param_values, observations)
}
\arguments{
\item{param_values}{\code{NumericVector} containing the probabilities of
a multinominal distribution. The length of this factor is the number of
categories - 1 since it contains only the parameters to be estimated.}

\item{observations}{\code{NumericVector} containing the number of
observations for each category of the multinominal distribution. The length
of this vector equals the number of categories.}
}
\value{
Returns the gradient as a \code{NumericVector}.
}
\description{
Function written in \code{C++} estimating the gradient of the log likelihood
function for a given parameter set and given observations.
}
\references{
Berding, Florian, and Pargmann, Julia (2022).Iota Reliability Concept
of the Second Generation.Measures for Content Analysis Done by
Humans or Artificial Intelligences. Berlin: Logos.
https://doi.org/10.30819/5581
}
