% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_new_rater.R
\docType{data}
\name{iotarelr_new_rater}
\alias{iotarelr_new_rater}
\title{Sample Vector}
\format{
A vector with the length of 318.
}
\usage{
iotarelr_new_rater
}
\description{
A vector containing the ratings of a new rater.
The data is not real and is only created for illustration purposes.
}
\keyword{datasets}
