% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_likelihood_multi_c}
\alias{log_likelihood_multi_c}
\title{Estimating log-likelihood in Condition Stage}
\usage{
log_likelihood_multi_c(probabilities, observations)
}
\arguments{
\item{probabilities}{\code{NumericVector} containing the probabilities of
a multinominal distribution.}

\item{observations}{\code{NumericVector} containing the number of
observations for each category of the multinominal distribution.}
}
\value{
Returns the log-likelihood as a single numeric value.
}
\description{
Function written in \code{C++} estimating the log-likelihood of a given
parameter set during the condition stage.
}
