% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iota_version_2.R
\name{sort_estimates}
\alias{sort_estimates}
\title{Sorting Estimates}
\usage{
sort_estimates(estimates)
}
\arguments{
\item{estimates}{list that contains the estimates generated by the function
\code{\link{EM_algo_c}}.}
}
\value{
Returns a list with sorted parameters.
}
\description{
Auxiliary function written in \code{R} for sorting the estimates generated by
the EM-Algorithm \code{\link{EM_algo_c}} according to the assumption of
weak superiority.
}
