% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iota_version_2.R
\name{compute_iota2}
\alias{compute_iota2}
\title{Computes Iota and its elements in version 2}
\usage{
compute_iota2(
  data,
  random_starts = 3,
  max_iterations = 5000,
  cr_rel_change = 1e-12,
  trace = TRUE
)
}
\arguments{
\item{data}{Data for which the elements should be estimated. Data must be
an object of type data.frame or matrix with cases in the rows and
raters in the columns.}

\item{random_starts}{An integer for the number of random starts for the
EM-algorithm.}

\item{max_iterations}{An integer for the maximum number of iterations within
the EM-algorithm.}

\item{cr_rel_change}{Positive numeric value for defining the convergence of the
EM-algorithm.}

\item{trace}{Boolean. \code{TRUE} if information on the estimation progress
should be printed to the console.}
}
\value{
Returns a \code{List} with the following three components:
The first component \code{estimates_categorical_level} comprises all
elements that describe the ratings on a categorical level. The elements are
sub-divided into raw estimates and chance-corrected estimates.
\describe{
\item{\code{raw_estimates}}{
\itemize{
\item{\code{alpha_reliability: }}{A vector containing the Alpha
Reliabilities for each category. These values represent probabilities.}
\item{\code{beta_reliability: }}{A vector containing the Beta Reliabilities for each
category. These values represent probabilities.}
\item{\code{assignment_error_matrix: }}{AEM-matrix containing the conditional
probabilities for assigning a unit of category i to categories 1 to n.}
}}
\item{\code{elements_chance_corrected}}{
\itemize{
\item{\code{alpha_reliability: }}
{A vector containing the chance corrected Alpha Reliabilities for each category.}
\item{\code{beta_reliability: }}
{A vector containing the chance corrected Beta Reliabilities for each category.}
\item{\code{iota: }}
{A vector containing the Iota values for each category.}
}}
}
The second component \code{estimates_scale_level} contains elements for
describing the quality of the ratings on a scale level. It comprises the
following elements:
\itemize{
\item{\code{avg_iota: }}
{Numeric value representing the mean of the Iota values.}
\item{\code{min_Iota: }}
{Numeric value representing the minimum of the Iota values.}
\item{\code{total_iota: }}
{Numeric value representing the weighted mean of the Iota values. Class sizes
are used for weighting.}
\item{\code{aem_index: }}
{Numeric value representing the deviation of the AEM from a matrix describing
random assignments.}
\item{\code{aem_index_weighted: }}
{Numeric value representing the deviation of the AEM from a matrix describing
random assignments. Categorical sizes are used for weighting. Nominal scale
level is assumed.}
\item{\code{aem_index_weighted_ord: }}
{Numeric value representing the deviation of the AEM from a matrix describing
random assignments. Categorical sizes are used for weighting. Ordinal scale
level is assumed.}
}
The third component \code{information} contains important information
regarding the parameter estimation. It comprises the following elements:
 \itemize{
\item{\code{log_likelihood: }}
{Log-likelihood of the used solution.}
\item{\code{Convergence: }}
{If estimation converged 0 else 1.}
\item{\code{categorical_sizes: }}
{Estimates categorical sizes. This is the estimated amount of the categories.}
\item{\code{conformity: }}
{\code{0} if the solution is in line with assumptions of weak superiority.
 A number greater 0 indicates the number of violations of the assumption
 of weak superiority.}.
\item{\code{n_condition: }}
{\code{TRUE} if the number of observed patterns equals or ist greater than
the degrees of freedom. \code{FALSE} if not}.
\item{\code{s_condition: }}
{\code{TRUE} if the rank of the derivates equals the degrees of freedoms.
 \code{FALSE} if not.}
\item{\code{deviation: }}
{Measure for describing the distance between the different solutions. 0
indicates that all solutions generated for the different start values
are the same.}
\item{\code{n_pattern: }}
{Numer of observed patterns.}
\item{\code{n_pattern_unordered: }}
{Numer of observed patterns if the order of codings are not considered.}
 \item{\code{n_parameter: }}
{Numer independently estimates parameters.}
 \item{\code{random_starts: }}
{Numer of random starts for the EM-algorithm.}
}
}
\description{
Computes all elements of the Iota Reliability Concept
}
\note{
This is an experimental version of an improved version of the iota
concept. Function and results may change in the future. We plan to implement
the final version at the beginng of 2023.
}
