% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writtenexams.R
\docType{data}
\name{iotarelr_written_exams}
\alias{iotarelr_written_exams}
\title{Example Data Set}
\format{
A data frame with 318 rows and 4 variables:
\describe{
  \item{Coder A}{Ratings of coder A.}
  \item{Coder B}{Ratings of coder B.}
  \item{Coder C}{Ratings of coder C.}
  \item{Gender}{Gender of the person who wrote the exam.}
}
}
\usage{
iotarelr_written_exams
}
\description{
A dataset containing the ratings of three coders for written exams.
It contains also the gender of the persons who wrote the exam. The data
is not real and is only created for illustrating purposes.
}
\keyword{datasets}
