% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{plot_iota}
\alias{plot_iota}
\title{Plot Iota}
\usage{
plot_iota(
  object,
  xlab = "Amount on all cases",
  ylab = "Categories",
  liota = "Assignment of the true category (Iota)",
  lcase2 = "Assignment to the false category",
  lcase3 = "Assignment from the false true category"
)
}
\arguments{
\item{object}{Estimates of Iota 2 created with \code{compute_iota2()},
\code{check_dgf} or \code{check_new_rater}.}

\item{xlab}{\code{Character} passed to xlab() from scale_fill_manual(). Label
of the x-axis.}

\item{ylab}{\code{Character} passed to ylab() from scale_fill_manual(). Label
of the y-axis.}

\item{liota}{\code{Character} passed to labels() from scale_fill_manual().
Label for Iota.Amount of cases that are assigned to the correct category.}

\item{lcase2}{\code{Character} passed to labels() from scale_fill_manual().
Label for the amount of cases that are assigned to a false category.}

\item{lcase3}{\code{Character} passed to labels() from scale_fill_manual().
Label for the amount of cases that are assigned from a false category.}
}
\value{
Function returns an object of class \code{gg, ggplot} illustrating how
the data of the different categories influence each other.
}
\description{
Function for creating a plot object that can be plotted via \link{ggplot2}.
}
\note{
An example for interpreting the plot can be found in:
}
